# goland_archpath/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для системы управления артефактами и заявками",
        "title": "ArchPath API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
        "/analysis-artifact-records/{request_id}/{artifact_id}": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Update an analysis artifact record (moderator only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis-artifact-records"
                ],
                "summary": "Update analysis artifact record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Artifact ID",
                        "name": "artifact_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/analysis_artifact_record.AnalysisArtifactRecord"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden: moderator access required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Update failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Delete an analysis artifact record (moderator only)",
                "tags": [
                    "analysis-artifact-records"
                ],
                "summary": "Delete analysis artifact record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Artifact ID",
                        "name": "artifact_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden: moderator access required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Deletion failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "analysis_artifact_record.AnalysisArtifactRecord": {
            "type": "object",
            "properties": {
                "artifact_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "is_main_entry": {
                    "type": "boolean"
                },
                "order": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}